+++ Kiosk/App/Views/Button Subclasses/Button.swift
@@ -1,4 +1,5 @@
 import UIKit
+import QuartzCore
 
 public class Button: ARFlatButton {
 
@@ -49,5 +50,23 @@ public class LargeKeypadButton: KeypadButton {
     public override func setup() {
         super.setup()
         self.titleLabel!.font = UIFont.sansSerifFontWithSize(20)
+    }
+}
+
+public class CircularBlackButton: ARBlackFlatButton {
+    public override func setup() {
+        super.setup()
+        RACObserve(self, "bounds").subscribeNext{ [weak self] (bounds) -> Void in
+            if let layer = self?.layer {
+                let width = CGRectGetWidth(layer.bounds)
+                let height = CGRectGetHeight(layer.bounds)
+                let smallestDimension = min(width, height)
+                layer.cornerRadius = smallestDimension / 2.0
+            }
+        }
+        
+        layer.borderWidth = 2
+        layer.borderColor = UIColor.whiteColor().CGColor
+        layer.masksToBounds = true
     }
 }