+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 * Bid Detail Previews are shown throughout - @orta
 * Use italics for title, and no italics for dates - @orta
 * Added a countdown on main page - @orta
+* Added help menu - @ashfurrow
 
 ### 0.0.4 - 9th Oct 2014